+++ include/rocksdb/perf_context.h
@@ -35,6 +35,9 @@ struct PerfContextByLevel {
   // total nanos spent on reading data from SST files
   uint64_t get_from_table_nanos;
 
+  uint64_t block_cache_hit_count = 0;     // total number of block cache hits
+  uint64_t block_cache_miss_count = 0;    // total number of block cache misses
+
   void Reset(); // reset all performance counters to zero
 };
 
@@ -205,6 +208,9 @@ struct PerfContext {
   uint64_t env_lock_file_nanos;
   uint64_t env_unlock_file_nanos;
   uint64_t env_new_logger_nanos;
+
+  uint64_t get_cpu_nanos;
+
   std::map<uint32_t, PerfContextByLevel>* level_to_perf_context;
   bool per_level_perf_context_enabled;
 };